+++ src/components/license.js
